requirements.txt
fastapi[standard]~=0.115
uvicorn[standard]~=0.32
pydantic-settings~=2.6
python-jose[cryptography]~=3.3
passlib[bcrypt]~=1.7
authlib~=1.3
httpx~=0.28
pytest
pytest-asyncio
